+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Date format: (YYYY-MM-DD)
   * The item matching implementation was updated to directly use Minecraft's underlying NBT comparison logic.
   * However, this also means that this logic is now part of the plugin's Minecraft version specific code that must be updated for every new version of Minecraft. A reflection based fallback implementation is available for Spigot servers, but is likely to break across server updates as well.
   * To more closely align with Minecraft's NBT matching logic used in commands and by the `custom_data` predicate, we also match partial lists now. Previously, we would not match partial lists to more closely align with how items are compared in villager trades. But in newer versions, the item comparison in villager trades has become more strict. And for the matching of config items, we prefer a more lenient partial matching of item data.
+* Config: Changes to the serialization of config item data:
+  * We no longer convert between legacy and alternative color codes. For recent Minecraft versions, the use of Minecraft's JSON text representation is recommended.
+  * Internal: Remove unused plain text serialization implementation.
 
 ## v2.23.10 (2025-07-27)
 ### Supported MC versions: 1.21.8, 1.21.7, 1.21.6, 1.21.5, 1.21.4, 1.21.3, 1.21.1, 1.21, 1.20.6